Delete utils file
